block: add logical block size emulation
Adds logical block size option to disk config so that a 4Kn disk can be
emulated to have a 512B logical block size. This enables booting with
UEFI from an OS on a 4Kn disk.

This is only functional for single-queue synchronous I/O on raw disk files,
i.e., num_queues=1,_disable_io_uring=on,_disable_aio=on

Signed-off-by: Sean Banko <sbanko@crusoeenergy.com>